Change in specification of printnames command
Removed the internal filtering process of the printnames command, which was separate from the user input values, to be able to output translations added by the mod. Since this increases the number of target translations by a large amount, the command argument was made variable length to improve filtering performance.